Organize project structure, separate color model.
