+++ lib/data/data.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:flutter/services.dart';
 import 'package:http/http.dart' as http;
 
-import 'models.dart';
+import '../domain/models.dart';
 
 class Api {
   const Api();
