+++ lib/domain/game.dart
@@ -1,10 +1,10 @@
 import 'dart:math';
 
-import 'package:capitals/data.dart';
+import 'package:capitals/data/data.dart';
 import 'package:flutter/material.dart';
 import 'package:palette_generator/palette_generator.dart';
 
-import '../models.dart';
+import 'models.dart';
 
 class GameLogic extends ChangeNotifier {
   static const _successGuess = 3;
@@ -137,11 +137,4 @@ class GameLogic extends ChangeNotifier {
     callback();
     notifyListeners();
   }
-}
-
-class ColorPair {
-  final Color main;
-  final Color second;
-
-  const ColorPair(this.main, this.second);
 }
